week7/flask-setup/requirements.txt
flask 
pandas
psycopg2-binary
pytest
python-dotenv